refactor: add ListKeys to CacheBackend interface, rename auto_sync to auto_pull
Move ListKeys from the separate ListableBackend interface into the core
CacheBackend interface. Move each implementation to its respective
backend file (fs.go, s3.go, gcs.go, azure.go, remote_wrapper.go).
Delete listable.go. Update mock backends in tests.

Also rename auto_sync -> auto_pull and autoSync -> autoPull throughout
config, CLI commands, and docs to match the pull command naming.
