+++ docs/src/content/docs/tracing/index.mdx
@@ -203,15 +203,15 @@ When using a remote cache backend (S3, GCS, Azure), traces written by other mach
 grog traces pull
 ```
 
-To automatically sync remote traces before every query command (`list`, `show`, `stats`, `export`), enable `auto_sync`:
+To automatically sync remote traces before every query command (`list`, `show`, `stats`, `export`), enable `auto_pull`:
 
 ```toml
 [traces]
 enabled = true
-auto_sync = true
+auto_pull = true
 ```
 
-With `auto_sync` enabled, running `grog traces list` will first download any new remote trace files, then query across all traces — both local and those from CI or other team members.
+With `auto_pull` enabled, running `grog traces list` will first download any new remote trace files, then query across all traces — both local and those from CI or other team members.
 
 ### Pruning old traces
 
@@ -243,7 +243,7 @@ There is no index file. DuckDB queries Parquet files directly via glob patterns,
 | Option | Type | Default | Description |
 |--------|------|---------|-------------|
 | `traces.enabled` | bool | `false` | Enable trace collection |
-| `traces.auto_sync` | bool | `false` | Automatically download remote traces before query commands |
+| `traces.auto_pull` | bool | `false` | Automatically download remote traces before query commands |
 | `traces.backend` | string | _(uses cache backend)_ | Override storage backend (`"s3"` or `"gcs"`) |
 | `traces.gcs.bucket` | string | | GCS bucket for trace storage |
 | `traces.gcs.prefix` | string | `/` | Prefix within the GCS bucket |
