+++ internal/caching/backends/cache_backend.go
@@ -25,6 +25,10 @@ type CacheBackend interface {
 
 	// Exists checks if a file exists in the cache with the given key.
 	Exists(ctx context.Context, path string, key string) (bool, error)
+
+	// ListKeys returns all keys under the given path that match the given suffix.
+	// Keys are returned as relative paths (e.g. "2026-03-30/trace-id.parquet").
+	ListKeys(ctx context.Context, path string, suffix string) ([]string, error)
 }
 
 func GetCacheBackend(
