+++ internal/config/config.go
@@ -235,7 +235,7 @@ type AzureCacheConfig struct {
 
 type TracesConfig struct {
 	Enabled  bool         `mapstructure:"enabled"`
-	AutoSync bool         `mapstructure:"auto_sync"`
+	AutoPull bool         `mapstructure:"auto_pull"`
 	Backend  CacheBackend `mapstructure:"backend"`
 	GCS      GCSCacheConfig `mapstructure:"gcs"`
 	S3       S3CacheConfig  `mapstructure:"s3"`
